requirements.txt
streamlit>=1.25
Pillow
ifcopenshell
pdf2image
Shapely
